+++ src/Edit.c
@@ -35,8 +35,6 @@
 #include "helpers.h"
 #include "resource.h"
 #include "EditHelper.h"
-#include "HLSelection.h"
-#include "StrToHex.h"
 
 extern HWND  hwndMain;
 extern HWND  hwndEdit;
@@ -4669,28 +4667,6 @@ void EditJumpTo(HWND hwnd, int iNewLine, int iNewCol)
   }
 }
 
-VOID	HL_Adjust_offset(int *pos, BOOL in)
-{
-#ifdef _DEBUG
-  HL_Trace("UTF8 %d", mEncoding[iEncoding].uFlags & NCP_UTF8);
-  HL_Trace("8Bit %d", mEncoding[iEncoding].uFlags & NCP_8BIT);
-  HL_Trace("UNicode %d", mEncoding[iEncoding].uFlags & NCP_UNICODE);
-  HL_Trace("offset is %d", *pos);
-#endif
-}
-
-void HL_Jump_offset(HWND hwnd, int iNewPos)
-{
-  HL_Adjust_offset(&iNewPos, TRUE);
-  SendMessage(hwnd, SCI_GOTOPOS, (WPARAM)iNewPos, 0);
-}
-
-void HL_Get_offset(HWND hwnd, int *out)
-{
-  *out = SendMessage(hwnd, SCI_GETCURRENTPOS, 0, 0);
-  HL_Adjust_offset(out, FALSE);
-}
-
 //=============================================================================
 //
 //  EditSelectEx()
@@ -5438,23 +5414,6 @@ HWND EditFindReplaceDlg(HWND hwnd, LPCEDITFINDREPLACE lpefr, BOOL bReplace)
   return hDlg;
 }
 
-int FindTextImpl(const HWND hwnd, const int searchFlags, struct TextToFind* pttf)
-{
-  return (int)SendMessage(hwnd, SCI_FINDTEXT, searchFlags, (LPARAM)pttf);
-}
-
-int FindTextTest(const HWND hwnd, const int searchFlags, const struct TextToFind* pttf, const int cpMin)
-{
-  struct TextToFind ttf = *pttf;
-  ttf.chrg.cpMin = cpMin;
-  return FindTextImpl(hwnd, searchFlags, &ttf);
-}
-
-BOOL CheckTextExists(const HWND hwnd, const int searchFlags, const struct TextToFind* pttf, const int iPos)
-{
-  return (FindTextTest(hwnd, searchFlags, pttf, iPos) >= 0);
-}
-
 //=============================================================================
 //
 //  EditFindNext()
@@ -5589,162 +5548,6 @@ BOOL EditFindPrev(HWND hwnd, LPCEDITFINDREPLACE lpefr, BOOL fExtendSelection)
     EditSelectEx(hwnd, max(iSelPos, iSelAnchor), ttf.chrgText.cpMin);
   }
   return TRUE;
-}
-
-void HL_Find_next_word(HWND hwnd, LPCEDITFINDREPLACE lpref, BOOL next)
-{
-  struct Sci_TextRange	tr;
-  struct Sci_TextToFind	ttf;
-  static char* szPrevWord = NULL;
-  int cpos, wlen, doclen, res, searchflags;
-  BOOL has;
-#define _HL_SEARCH_FOR_WORD_LIMIT 0x100
-  HL_TRACE(L"look for next(%d) word", next);
-  ZeroMemory(&ttf, sizeof(ttf));
-  ttf.lpstrText = 0;
-  cpos = SendMessage(hwnd, SCI_GETCURRENTPOS, 0, 0);
-  doclen = SendMessage(hwnd, SCI_GETTEXTLENGTH, 0, 0);
-  tr.chrg.cpMin = SendMessage(hwnd, SCI_WORDSTARTPOSITION, cpos, TRUE);
-  tr.chrg.cpMax = SendMessage(hwnd, SCI_WORDENDPOSITION, cpos, TRUE);
-  wlen = tr.chrg.cpMax - tr.chrg.cpMin;
-  res = 0;
-
-  tr.lpstrText = HL_Alloc(wlen + 1);
-  SendMessage(hwnd, SCI_GETTEXTRANGE, 0, (LPARAM)&tr);
-
-  const int iSelCount = (int)SendMessage(hwnd, SCI_GETSELECTIONEND, 0, 0) -
-                        (int)SendMessage(hwnd, SCI_GETSELECTIONSTART, 0, 0);
-
-  HL_Free(tr.lpstrText);
-
-  if (iSelCount > 0)
-  {
-    const size_t prevWordLength = szPrevWord ? strlen(szPrevWord) + 1 : 0;
-    if (szPrevWord && (prevWordLength > 0))
-    {
-      tr.lpstrText = HL_Alloc(prevWordLength);
-      lstrcpynA(tr.lpstrText, szPrevWord, prevWordLength);
-      ttf.lpstrText = tr.lpstrText;
-      res = 1;
-    }
-  }
-  if (res == 0)
-  {
-    has = wlen > 0;
-
-    // look up for new word for search
-    if (!has)
-    {
-      tr.chrg.cpMin = next ? cpos : max(cpos - _HL_SEARCH_FOR_WORD_LIMIT, 0);
-      tr.chrg.cpMax = next ? min(cpos + _HL_SEARCH_FOR_WORD_LIMIT, doclen) : cpos;
-      wlen = tr.chrg.cpMax - tr.chrg.cpMin;
-      if (wlen > 0)
-      {
-        int counter;
-        char symb;
-        //
-        tr.lpstrText = HL_Alloc(wlen + 1);
-        SendMessage(hwnd, SCI_GETTEXTRANGE, 0, (LPARAM)&tr);
-        counter = 0;
-        while (counter <= wlen)
-        {
-          ++counter;
-          symb = next ? tr.lpstrText[counter] : tr.lpstrText[wlen - counter];
-          if (HL_IS_LITERAL(symb))
-          {
-            if (!res)
-            {
-              res = counter;
-            }
-          }
-          else
-          {
-            if (res)
-            {
-              if (next)
-              {
-                tr.chrg.cpMax = cpos + counter;
-                tr.lpstrText[counter] = '\0';
-                ttf.lpstrText = tr.lpstrText + res;
-              }
-              else
-              {
-                tr.chrg.cpMin = cpos - res;
-                tr.lpstrText[wlen - res + 1] = '\0';
-                ttf.lpstrText = tr.lpstrText + wlen - counter + 1;
-              }
-              break;
-            }
-          }
-        }
-      }
-    }
-    else
-    {
-      tr.lpstrText = HL_Alloc(wlen + 1);
-      SendMessage(hwnd, SCI_GETTEXTRANGE, 0, (LPARAM)&tr);
-      ttf.lpstrText = tr.lpstrText;
-      res = 1;
-    }
-  }
-  //
-  if (res)
-  {
-    HL_TRACE("search for '%s' ", ttf.lpstrText);
-    if (next)
-    {
-      ttf.chrg.cpMin = tr.chrg.cpMax;
-      ttf.chrg.cpMax = doclen;
-    }
-    else
-    {
-      ttf.chrg.cpMin = tr.chrg.cpMin;
-      ttf.chrg.cpMax = 0;
-    }
-    searchflags = SCFIND_WHOLEWORD;
-    if (iFindWordMatchCase != 0)
-      searchflags |= SCFIND_MATCHCASE;
-
-    res = FindTextImpl(hwnd, searchflags, &ttf);
-    const BOOL bTextFound = (res >= 0);
-    UpdateFindIcon(bTextFound && (FindTextTest(hwnd, searchflags, &ttf, res + 1) >= 0));
-
-    if ((-1 == res) && (iFindWordWrapAround != 0))
-    {
-      if (next)
-      {
-        ttf.chrg.cpMin = 0;
-        ttf.chrg.cpMax = tr.chrg.cpMin;
-      }
-      else
-      {
-        ttf.chrg.cpMin = doclen;
-        ttf.chrg.cpMax = tr.chrg.cpMax;
-      }
-      res = FindTextImpl(hwnd, searchflags, &ttf);
-      UpdateFindIcon(res >= 0);
-    }
-    if (res >= 0)
-    {
-      EditSelectEx(hwnd, ttf.chrgText.cpMin, ttf.chrgText.cpMax);
-    }
-    else
-    {
-      SendMessage(hwnd, SCI_SETCURRENTPOS, cpos, 0);
-    }
-    if (ttf.lpstrText)
-    {
-      const char* lpstrText = ttf.lpstrText;
-      HL_Free(szPrevWord);
-      szPrevWord = HL_Alloc(strlen(lpstrText) + 1);
-      lstrcpynA(szPrevWord, lpstrText, strlen(lpstrText) + 1);
-    }
-    if (tr.lpstrText)
-    {
-      HL_Free(tr.lpstrText);
-      tr.lpstrText = 0;
-    }
-  }
 }
 
 //=============================================================================
@@ -7435,304 +7238,6 @@ int FileVars_GetEncoding(LPFILEVARS lpfv)
   {
     return (-1);
   }
-}
-
-BOOL HL_Open_nextFs_file(HWND hwnd, LPCWSTR file, BOOL next)
-{
-  WCHAR	dirname[MAX_PATH], odn[MAX_PATH], found_path[MAX_PATH], *filename;
-  HANDLE	hFind = INVALID_HANDLE_VALUE;
-  WIN32_FIND_DATA	ffd;
-  INT		cmp_res;
-  *found_path = L'\0';
-  filename = PathFindFileName(file);
-  StrCpy(dirname, file);
-  if (!PathRemoveFileSpec(dirname))
-  {
-    return FALSE;
-  }
-  if (L'\\' != dirname[lstrlen(dirname) - 1])
-  {
-    StrCat(dirname, L"\\");
-  }
-  StrCpy(odn, dirname);
-  StrCat(dirname, L"*");
-  hFind = FindFirstFile(dirname, &ffd);
-  if (INVALID_HANDLE_VALUE == hFind)
-  {
-    return FALSE;
-  }
-#define _HL_COMPARE_FILES( F1 , F2 )  (HL_Compare_files(F1,F2))
-  do
-  {
-    if (0 == (ffd.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY))
-    {
-      cmp_res = _HL_COMPARE_FILES(filename, ffd.cFileName);
-      HL_TRACE(L"%S vs %S = %d", ffd.cFileName, filename, cmp_res);
-      if ((next && cmp_res >= 0) || (!next&&cmp_res <= 0))
-      {
-        continue;
-      }
-      if (*found_path)
-      {
-        cmp_res = _HL_COMPARE_FILES(found_path, ffd.cFileName);
-      }
-      else
-      {
-        cmp_res = 0;
-      }
-      HL_TRACE(L"%S vs %S = %d", ffd.cFileName, found_path, cmp_res);
-      if ((next && cmp_res >= 0) || (!next&&cmp_res <= 0))
-      {
-        StrCpy(found_path, ffd.cFileName);
-        HL_TRACE(L"saved %S", found_path);
-      }
-    }
-  } while (FindNextFile(hFind, &ffd));
-  FindClose(hFind);
-  if (*found_path)
-  {
-    StrCat(odn, found_path);
-    HL_TRACE(L"file to open %S", odn);
-    FileLoad(FALSE, FALSE, FALSE, FALSE, odn);
-  }
-  return TRUE;
-}
-
-void HL_Unwrap_selection(HWND hwnd, BOOL quote_mode)
-{
-  int cpos, len, temp, pos_left, pos_right;
-  struct Sci_TextRange tr_1, tr_2;
-  BOOL found;
-  const static int max_region_to_scan = 1024;
-  const static int max_brackets_to_skip = 100;
-  cpos = SendMessage(hwnd, SCI_GETCURRENTPOS, 0, 0);
-  len = SendMessage(hwnd, SCI_GETTEXTLENGTH, 0, 0);
-  tr_1.chrg.cpMax = cpos;
-  tr_1.chrg.cpMin = max(0, cpos - max_region_to_scan);
-  tr_1.lpstrText = NULL;
-  tr_2.chrg.cpMin = cpos;
-  tr_2.chrg.cpMax = min(len, cpos + max_region_to_scan);
-  tr_2.lpstrText = NULL;
-  {
-    temp = abs(tr_1.chrg.cpMax - tr_1.chrg.cpMin);
-    if (!temp) goto OUT_OF_UNWRAP;
-    tr_1.lpstrText = HL_Alloc(temp + 1);
-    SendMessage(hwnd, SCI_GETTEXTRANGE, 0, (LPARAM)&tr_1);
-  }
-  {
-    temp = abs(tr_2.chrg.cpMax - tr_2.chrg.cpMin);
-    if (!temp) goto OUT_OF_UNWRAP;
-    tr_2.lpstrText = HL_Alloc(temp + 1);
-    SendMessage(hwnd, SCI_GETTEXTRANGE, 0, (LPARAM)&tr_2);
-  }
-  pos_left = tr_1.chrg.cpMax, pos_right = tr_2.chrg.cpMin;
-  found = FALSE;
-  if (quote_mode)
-  {
-    const char* _quotes = "\"'`";
-    char* qchl = NULL;
-    // search left
-    while (1)
-    {
-      char lch = tr_1.lpstrText[pos_left - tr_1.chrg.cpMin - 1];
-      if (lch)
-      {
-        if (qchl = strchr(_quotes, lch))
-        {
-          HL_TRACE("Left quote found '%c'", lch);
-          break;
-        }
-      }
-      if (--pos_left <= tr_1.chrg.cpMin)
-      {
-        qchl = NULL;
-        break;
-      }
-    }
-
-    // go right
-    while (qchl)
-    {
-      char rch = tr_2.lpstrText[pos_right - tr_2.chrg.cpMin];
-      if (rch)
-      {
-        if (rch == *qchl)
-        {
-          HL_TRACE("Right quote found '%c'", rch);
-          break;
-        }
-      }
-      if (++pos_right > tr_2.chrg.cpMax)
-      {
-        qchl = NULL;
-        break;
-      }
-    }
-    found = NULL != qchl;
-  }
-  else
-  {
-    const char* _left_braces = "<{([";
-    const char* _right_braces = ">})]";
-    char* tchl = NULL, *tchr = NULL, *qchl = NULL;
-    int	  skipcl = 0, skipcr = 0;
-    int*  skipl = HL_Alloc(max_brackets_to_skip * sizeof(int));
-
-    // search left
-  RESUME_SEARCH:
-    while (1)
-    {
-      char lch = tr_1.lpstrText[pos_left - tr_1.chrg.cpMin - 1];
-      if (lch)
-      {
-        if (tchl = strchr(_left_braces, lch))
-        {
-          if (skipcl)
-          {
-            int ti = 0;
-            for (; ti < skipcl; ++ti)
-            {
-              if (tchl - _left_braces == skipl[ti])
-              {
-                HL_TRACE("Skipped braces pair found '%c'", *tchl);
-                skipl[ti] = -1;
-                goto NEXT;
-              }
-            }
-          }
-          {
-            HL_TRACE("Left bracket found '%c'", lch);
-            break;
-          }
-        }
-        if (tchl = strchr(_right_braces, lch))
-        {
-          skipl[skipcl++] = tchl - _right_braces;
-        }
-      }
-    NEXT:
-      if (--pos_left <= tr_1.chrg.cpMin)
-      {
-        tchl = NULL;
-        break;
-      }
-    }
-    // go right
-    while (tchl)
-    {
-      char rch = tr_2.lpstrText[pos_right - tr_2.chrg.cpMin];
-      if (rch)
-      {
-        if (tchr = strchr(_right_braces, rch))
-        {
-          if (tchr - _right_braces == tchl - _left_braces)
-          {
-            if (skipcr)
-            {
-              HL_TRACE("Skip right bracket '%c' (%d to skip)", rch, skipcr);
-              --skipcr;
-            }
-            else
-            {
-              HL_TRACE("Right bracket found '%c'", rch);
-              break;
-            }
-          }
-          else
-          {
-            tchr = NULL;
-            HL_TRACE("Bad right bracket found '%c'", rch);
-          }
-        }
-        if (tchr = strchr(_left_braces, rch))
-        {
-          if (tchr == tchl)
-          {
-            ++skipcr;
-          }
-        }
-      }
-      if (++pos_right > tr_2.chrg.cpMax)
-      {
-        tchr = NULL;
-        break;
-      }
-    }
-    if (tchl && !tchr && --pos_left > tr_1.chrg.cpMin)
-    {
-      tchl = NULL;
-      tchr = NULL;
-      skipcr = 0;
-      pos_right = tr_2.chrg.cpMin;
-      goto RESUME_SEARCH;
-    }
-    HL_Free(skipl);
-    found = tchr && tchl;
-  }
-  // remove
-  if (found)
-  {
-    HL_TRACE("removing braces OR quotes at %d and %d", pos_left - 1, pos_right);
-    SendMessage(hwnd, SCI_BEGINUNDOACTION, 0, 0);
-    SendMessage(hwnd, SCI_DELETERANGE, pos_left - 1, 1);
-    SendMessage(hwnd, SCI_DELETERANGE, pos_right - 1 /*remember offset from prev line*/, 1);
-    SendMessage(hwnd, SCI_SETSEL, pos_left - 1, pos_right - 1);
-    SendMessage(hwnd, SCI_ENDUNDOACTION, 0, 0);
-  }
-OUT_OF_UNWRAP:
-  HL_Free(tr_1.lpstrText);
-  HL_Free(tr_2.lpstrText);
-}
-
-void HL_Escape_html(HWND hwnd)
-{
-  int beg, end, res;
-  size_t symb;
-  BOOL changed;
-  struct Sci_TextToFind ttf;
-  const char* _source = "&<>";
-  const char* _target[] = { "&amp;", "&lt;", "&gt;" };
-  assert(strlen(_source) == COUNTOF(_target));
-  SendMessage(hwnd, SCI_BEGINUNDOACTION, 0, 0);
-  beg = SendMessage(hwnd, SCI_GETSELECTIONSTART, 0, 0);
-  end = SendMessage(hwnd, SCI_GETSELECTIONEND, 0, 0);
-  if (beg == end)
-  {
-    beg = 0;
-    end = SendMessage(hwnd, SCI_GETTEXTLENGTH, 0, 0);
-  }
-  ttf.lpstrText = HL_Alloc(2);
-  ttf.lpstrText[1] = '\0';
-  changed = FALSE;
-  for (symb = 0; symb < strlen(_source); ++symb)
-  {
-    ttf.chrg.cpMin = beg;
-    ttf.chrg.cpMax = end;
-    ttf.lpstrText[0] = _source[symb];
-    res = 0;
-    while (-1 != res)
-    {
-      res = SendMessage(hwnd, SCI_FINDTEXT, 0, (LPARAM)&ttf);
-      if (-1 != res)
-      {
-          {
-            assert(ttf.chrgText.cpMax == ttf.chrgText.cpMin + 1);
-            SendMessage(hwnd, SCI_DELETERANGE, ttf.chrgText.cpMin, 1);
-            SendMessage(hwnd, SCI_INSERTTEXT, ttf.chrgText.cpMin, (LPARAM)_target[symb]);
-            ttf.chrg.cpMin = ttf.chrgText.cpMax;
-            end += strlen(_target[symb]) - 1;
-            ttf.chrg.cpMax = end;
-            changed = TRUE;
-          }
-        }
-      }
-    }
-  if (changed)
-  {
-    SendMessage(hwnd, SCI_SETSEL, beg, beg);
-  }
-  HL_Free(ttf.lpstrText);
-  SendMessage(hwnd, SCI_ENDUNDOACTION, 0, 0);
 }
 
 ///   End of Edit.c   \\\
